+++ MiniProjet.Core/Models/Brand.cs
@@ -9,7 +9,7 @@ namespace MiniProjet.Core.Models
 {
     public class Brand
     {
-        public Guid BrandId { get; set; }
+        public Guid BrandId { get; private set; } = Guid.NewGuid();
 
         public String  BrandName { get; set; }
