+++ MiniProjet.Core/Repositories/ProductRepo/ProductRepository.cs
@@ -1,4 +1,5 @@
 ﻿using MiniProjet.Core.Models;
+using MiniProjet.Core.Repositories.BrandRepo;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,9 +10,11 @@ namespace MiniProjet.Core.Repositories.ProductRepo
 {
     public class ProductRepository : IProductRepo
     {
+        private IBrandRepo brandRepo;
         private List<Product> Products;
-        public ProductRepository(List<Product> products)
+        public ProductRepository(List<Product> products,IBrandRepo _brandRepo)
         {
+            brandRepo = _brandRepo;
             Products = products;
         }
 
@@ -28,8 +31,22 @@ namespace MiniProjet.Core.Repositories.ProductRepo
 
         public void AddNewProduct(Product product)
         {
-            if (!ValidateProduct(product)) Products.Add(product);
+            Guid guid = Guid.NewGuid();
+            if (!ValidateProduct(product))
+            {
+                Products.Add(product);
+
+                brandRepo.AddBrands(product.Brand);
+                var brand = brandRepo.GetBrand(product.Brand.BrandName);
+                brand.Products.Add(product);
+            }
+               
+            
+          
         }
+
+
+
 
         public Product GetProduct(Guid IdProduct)
         {
@@ -67,7 +84,7 @@ namespace MiniProjet.Core.Repositories.ProductRepo
 
         public bool ValidateProduct(Product product)
         {
-            return   string.IsNullOrEmpty(product.ProductName) || product.ProductPrice <= 0 ;
+            return   string.IsNullOrEmpty(product.ProductName) || product.ProductPrice <= 0  || product.Brand ==null;
         }
     }
 }
